# baseline-db/openapi.yml
openapi: 3.0.0
info:
  title: merkle-tree
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:9000
    variables: {}
paths:
  /leaf:
    post:
      tags:
        - leaf
      summary: /leaf -> inserts a leaf object
      description: inserts a leaf object
      operationId: /leaf->insertsaleafobject
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1leaf-%3Einsertsaleafobjectrequest'
            example:
              contractName: MerkleTreeControllerSHA
              leaf:
                value: 0xabc123-7
                leafIndex: 2
                blockNumber: 60000000
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /leaves:
    post:
      tags:
        - leaf
      summary: /leaves -> inserts many leaf objects
      description: >-
        {
        		"value": "0xabc123-0",
        		"leafIndex": 0,
        		"blockNumber": 60000000
        	},
        	{
        		"value": "0xabc123-2",
        		"leafIndex": 2,
        		"blockNumber": 60000003
        	},
        	{
        		"value": "0xabc123-3",
        		"leafIndex": 3,
        		"blockNumber": 60000000
        	},
        		{
        		"value": "0xabc123-2",
        		"leafIndex": 7,
        		"blockNumber": 60000003
        	},
        	{
        		"value": "0xabc123-3",
        		"leafIndex": 8,
        		"blockNumber": 60000000
        	},
      operationId: /leaves->insertsmanyleafobjects
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1leaves-%3Einsertsmanyleafobjectsrequest'
            example:
              contractName: MerkleTreeControllerSHA
              leaves:
                - value: 0xabc123-3
                  leafIndex: 1
                  blockNumber: 60000000
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - leaf
      summary: /leaves -> get leaves by array of indices
      operationId: /leaves->getleavesbyarrayofindices
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
...
  /leaf/index:
    get:
      tags:
        - leaf
      summary: /leaf/index -> gets a leaf at leafIndex
      operationId: /leaf/index->getsaleafatleafIndex
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /leaf/value:
    get:
      tags:
        - leaf
      summary: /leaf/value -> gets a leaf with a given value
      operationId: /leaf/value->getsaleafwithagivenvalue
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /leaves/check:
    get:
      tags:
        - leaf
      summary: /leaves/check -> check for missing leaves
      operationId: /leaves/check->checkformissingleaves
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /leaves/count:
    get:
      tags:
        - leaf
      summary: /leaves/count -> count the number of leaves in the db
      operationId: /leaves/count->countthenumberofleavesinthedb
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
...
  /start:
    post:
      tags:
        - merkle-tree
      summary: /start -> starts the event filter
      description: inserts a node object
      operationId: /start->startstheeventfilter
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1start-%3Estartstheeventfilterrequest'
            example:
              contractName: MerkleTreeControllerSHA
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
...
  /update:
    patch:
      tags:
        - merkle-tree
      summary: /update -> updates the merkle tree
      operationId: /update->updatesthemerkletree
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1update-%3Eupdatesthemerkletreerequest'
            example:
              contractName: MerkleTreeControllerSHA
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
...